Modified the main file so that the user can decide the secret

+++ src/fr.c
@@ -42,7 +42,7 @@ uint64_t measure_access_time(const uint8_t *addr) {
 /// Flush+Reload implementation
 FlushReload fr_init(void) {
     uint8_t *buffer = malloc(N_BYTE_VALUES * PAGE_SIZE * sizeof(uint8_t));
-    for (int i = 0; i < N_BYTE_VALUES * PAGE_SIZE; i++) buffer[i] = 219;
+    for (size_t i = 0; i < N_BYTE_VALUES * PAGE_SIZE; i++) buffer[i] = 219;
     return (FlushReload){._buffer = buffer};
 }
 
@@ -52,7 +52,7 @@ void fr_free(FlushReload *fr) {
 }
 
 void fr_reset(const FlushReload *fr) {
-    for (int page_start = 0; page_start < N_BYTE_VALUES * PAGE_SIZE; page_start += PAGE_SIZE) {
+    for (size_t page_start = 0; page_start < N_BYTE_VALUES * PAGE_SIZE; page_start += PAGE_SIZE) {
         flush(fr->_buffer + page_start);
     }
 }
@@ -65,7 +65,7 @@ uint8_t fr_get(const FlushReload *fr) {
     uint8_t byte_index_with_min_access_time = 0;
     uint64_t min_access_time = 1 << 31;
 
-    for (int i = 0; i < N_BYTE_VALUES; i++) {
+    for (size_t i = 0; i < N_BYTE_VALUES; i++) {
         const uint8_t byte_idx = (113 * i + 1) % N_BYTE_VALUES; // pseudo-randomness
 
         const uint64_t access_time = measure_access_time(fr->_buffer + byte_idx * PAGE_SIZE);
